refactor(cf): relocate cf admin signal configs out of list framework
