+++ src/frontend/packages/cloud-foundry/src/shared/signal-list-configs/cf-feature-flags/cf-feature-flags-signal-config.service.ts
@@ -3,54 +3,52 @@ import { HttpClient } from '@angular/common/http';
 
 import { ListStateStore } from '@stratosui/core';
 
-import { CnsiStacksSource } from '../../../../../services/data-sources/cnsi-stacks-source';
-import { ViewPipeline, SortSpec } from '../../../../../services/data-sources/view-pipeline';
-import type { StStack } from '../../../../../services/endpoint-data/stratos-types';
+import { CnsiFeatureFlagsSource } from '../../../services/data-sources/cnsi-feature-flags-source';
+import { ViewPipeline, SortSpec } from '../../../services/data-sources/view-pipeline';
+import type { StFeatureFlag } from '../../../services/endpoint-data/stratos-types';
 
-// CF Stacks list config — single-CNSI, read-only. Drives the per-CF
-// /cloud-foundry/:cnsi/stacks tab. Stacks are foundation-level metadata
-// (rootfs flavors); they don't relate to orgs / spaces / apps and are
-// rarely many (typically <10), so this service stays simple: own the
-// fetch, expose a stacks() signal, run it through ViewPipeline. No
-// writes — stack creation/edit/delete are platform-admin operations
-// not surfaced here.
+// CF Feature Flags list config — single-CNSI, read-only. Drives the
+// per-CF /cloud-foundry/:cnsi/feature-flags tab. Foundations expose
+// ~15 flags so this service stays simple: own the fetch, expose a
+// featureFlags() signal, run it through ViewPipeline. No writes —
+// flag toggle is a platform-admin operation not surfaced here.
 @Injectable({ providedIn: 'root' })
-export class CfStacksSignalConfigService {
+export class CfFeatureFlagsSignalConfigService {
   private readonly http = inject(HttpClient);
   private readonly injector = inject(Injector);
 
   private cnsiGuid = '';
-  private source?: CnsiStacksSource;
+  private source?: CnsiFeatureFlagsSource;
 
-  private readonly state = inject(ListStateStore).bind('cf-stacks', {
+  private readonly state = inject(ListStateStore).bind('cf-feature-flags', {
     viewMode: 'card',
     pageSize: [24, 25],
     pageIndex: [0, 0],
     sort: [{ field: 'name', direction: 'asc' }, { field: 'name', direction: 'asc' }],
   });
 
-  readonly filter: WritableSignal<(s: StStack) => boolean> = signal(() => true);
-  readonly sort = this.state.sort as WritableSignal<SortSpec<StStack>>;
+  readonly filter: WritableSignal<(f: StFeatureFlag) => boolean> = signal(() => true);
+  readonly sort = this.state.sort as WritableSignal<SortSpec<StFeatureFlag>>;
   readonly pageSize = this.state.pageSize;
   readonly pageIndex = this.state.pageIndex;
   readonly nameFilter: WritableSignal<string> = signal('');
   readonly viewMode = this.state.viewMode;
 
-  private readonly _stacks: WritableSignal<StStack[]> = signal([]);
-  readonly stacks: Signal<StStack[]> = this._stacks.asReadonly();
+  private readonly _featureFlags: WritableSignal<StFeatureFlag[]> = signal([]);
+  readonly featureFlags: Signal<StFeatureFlag[]> = this._featureFlags.asReadonly();
 
   private readonly _hasLoadedOnce: WritableSignal<boolean> = signal(false);
   readonly hasLoadedOnce: Signal<boolean> = this._hasLoadedOnce.asReadonly();
 
-  private readonly _sortExtractors: WritableSignal<Map<string, (row: StStack) => unknown>> = signal(new Map());
+  private readonly _sortExtractors: WritableSignal<Map<string, (row: StFeatureFlag) => unknown>> = signal(new Map());
 
-  view!: ViewPipeline<StStack>;
+  view!: ViewPipeline<StFeatureFlag>;
 
   initialize(cnsiGuid: string): void {
     this.cnsiGuid = cnsiGuid;
-    this.source = new CnsiStacksSource(cnsiGuid, this.http);
-    this.view = new ViewPipeline<StStack>(
-      this.stacks,
+    this.source = new CnsiFeatureFlagsSource(cnsiGuid, this.http);
+    this.view = new ViewPipeline<StFeatureFlag>(
+      this.featureFlags,
       this.filter,
       this.sort,
       this.pageSize,
@@ -61,9 +59,9 @@ export class CfStacksSignalConfigService {
     runInInjectionContext(this.injector, () => {
       effect(() => {
         const q = this.nameFilter().trim().toLowerCase();
-        this.filter.set((s: StStack) => {
+        this.filter.set((f: StFeatureFlag) => {
           if (!q) return true;
-          return (s.name ?? '').toLowerCase().includes(q);
+          return (f.name ?? '').toLowerCase().includes(q);
         });
       });
     });
@@ -72,14 +70,14 @@ export class CfStacksSignalConfigService {
   async loadAll(): Promise<void> {
     if (!this.source) return;
     await this.source.load();
-    this._stacks.set([...this.source.items()]);
+    this._featureFlags.set([...this.source.items()]);
     this._hasLoadedOnce.set(true);
   }
 
   async refresh(): Promise<void> {
     if (!this.source) return;
     await this.source.refresh();
-    this._stacks.set([...this.source.items()]);
+    this._featureFlags.set([...this.source.items()]);
     this._hasLoadedOnce.set(true);
   }
 
@@ -89,7 +87,7 @@ export class CfStacksSignalConfigService {
     this.pageIndex.set(0);
   }
 
-  registerSortExtractor(fieldKey: string, extractor: (row: StStack) => unknown): void {
+  registerSortExtractor(fieldKey: string, extractor: (row: StFeatureFlag) => unknown): void {
     this._sortExtractors.update(curr => {
       const next = new Map(curr);
       next.set(fieldKey, extractor);
