+++ src/frontend/packages/cloud-foundry/src/shared/signal-list-configs/cf-quotas/cf-org-quotas-signal-config.service.spec.ts
@@ -3,18 +3,24 @@ import { HttpClient } from '@angular/common/http';
 import { of } from 'rxjs';
 import { TestBed } from '@angular/core/testing';
 import { provideZonelessChangeDetection } from '@angular/core';
-import { CfSecurityGroupsSignalConfigService } from './cf-security-groups-signal-config.service';
-import type { StSecurityGroup } from '../../../../../services/endpoint-data/stratos-types';
+import { CfOrgQuotasSignalConfigService } from './cf-org-quotas-signal-config.service';
+import type { StOrgQuota } from '../../../services/endpoint-data/stratos-types';
 
-function makeGroup(overrides: Partial<StSecurityGroup>): StSecurityGroup {
+function makeQuota(overrides: Partial<StOrgQuota>): StOrgQuota {
   return {
-    guid: 'sg-1',
-    name: 'public_networks',
-    globallyEnabledRunning: true,
-    globallyEnabledStaging: false,
-    ruleCount: 2,
-    runningSpaceCount: 1,
-    stagingSpaceCount: 0,
+    guid: 'q-1',
+    name: 'default',
+    totalMemoryInMB: 102400,
+    totalInstanceMemoryInMB: 2048,
+    totalInstances: 1000,
+    totalAppTasks: 500,
+    paidServicesAllowed: true,
+    totalServiceInstances: 250,
+    totalServiceKeys: 250,
+    totalRoutes: 1000,
+    totalReservedPorts: 100,
+    totalDomains: 10,
+    organizationCount: 2,
     cnsiGuid: 'cnsi-1',
     createdAt: '2026-04-22T12:00:00Z',
     updatedAt: '2026-04-22T12:00:00Z',
@@ -22,12 +28,12 @@ function makeGroup(overrides: Partial<StSecurityGroup>): StSecurityGroup {
   };
 }
 
-function makeHttp(groups: StSecurityGroup[]): HttpClient {
+function makeHttp(quotas: StOrgQuota[]): HttpClient {
   return {
     get: vi.fn(() => of({
-      resources: groups,
+      resources: quotas,
       pagination: {
-        totalResults: groups.length,
+        totalResults: quotas.length,
         totalPages: 1,
         next: null,
         previous: null,
@@ -38,26 +44,26 @@ function makeHttp(groups: StSecurityGroup[]): HttpClient {
   } as unknown as HttpClient;
 }
 
-function makeSvc(http: HttpClient): CfSecurityGroupsSignalConfigService {
+function makeSvc(http: HttpClient): CfOrgQuotasSignalConfigService {
   TestBed.configureTestingModule({
     providers: [
       provideZonelessChangeDetection(),
       { provide: HttpClient, useValue: http },
-      CfSecurityGroupsSignalConfigService,
+      CfOrgQuotasSignalConfigService,
     ],
   });
-  return TestBed.inject(CfSecurityGroupsSignalConfigService);
+  return TestBed.inject(CfOrgQuotasSignalConfigService);
 }
 
 beforeEach(() => TestBed.resetTestingModule());
 
-describe('CfSecurityGroupsSignalConfigService', () => {
-  it('exposes empty groups before initialize', () => {
+describe('CfOrgQuotasSignalConfigService', () => {
+  it('exposes empty quotas before initialize', () => {
     const svc = makeSvc(makeHttp([]));
-    expect(svc.securityGroups()).toEqual([]);
+    expect(svc.orgQuotas()).toEqual([]);
   });
 
-  it('exposes the filter, sort, pageSize, pageIndex, nameFilter signals', () => {
+  it('exposes filter, sort, pageSize, pageIndex, nameFilter signals', () => {
     const svc = makeSvc(makeHttp([]));
     svc.initialize('cnsi-1');
     expect(svc.filter).toBeDefined();
@@ -67,25 +73,25 @@ describe('CfSecurityGroupsSignalConfigService', () => {
     expect(svc.nameFilter).toBeDefined();
   });
 
-  it('builds a ViewPipeline driven by the securityGroups signal', () => {
+  it('builds a ViewPipeline driven by the orgQuotas signal', () => {
     const svc = makeSvc(makeHttp([]));
     svc.initialize('cnsi-1');
     expect(svc.view).toBeDefined();
   });
 
-  it('loads security groups from /pp/v1/cf/security_groups/:cnsi', async () => {
+  it('loads org quotas from /pp/v1/cf/organization_quotas/:cnsi', async () => {
     const http = makeHttp([
-      makeGroup({ guid: 'sg-1', name: 'public_networks' }),
-      makeGroup({ guid: 'sg-2', name: 'dns', globallyEnabledStaging: true }),
+      makeQuota({ guid: 'q-1', name: 'default' }),
+      makeQuota({ guid: 'q-2', name: 'large' }),
     ]);
     const svc = makeSvc(http);
     svc.initialize('cnsi-1');
     await svc.loadAll();
 
-    expect(svc.securityGroups().length).toBe(2);
-    expect(svc.securityGroups()[0].name).toBe('public_networks');
+    expect(svc.orgQuotas().length).toBe(2);
+    expect(svc.orgQuotas()[0].name).toBe('default');
     expect(http.get).toHaveBeenCalledWith(
-      expect.stringContaining('/pp/v1/cf/security_groups/cnsi-1'),
+      expect.stringContaining('/pp/v1/cf/organization_quotas/cnsi-1'),
     );
   });
