+++ src/frontend/packages/cloud-foundry/src/shared/signal-list-configs/cf-stacks/cf-stacks-signal-config.service.spec.ts
@@ -3,24 +3,15 @@ import { HttpClient } from '@angular/common/http';
 import { of } from 'rxjs';
 import { TestBed } from '@angular/core/testing';
 import { provideZonelessChangeDetection } from '@angular/core';
-import { CfOrgQuotasSignalConfigService } from './cf-org-quotas-signal-config.service';
-import type { StOrgQuota } from '../../../../../services/endpoint-data/stratos-types';
+import { CfStacksSignalConfigService } from './cf-stacks-signal-config.service';
+import type { StStack } from '../../../services/endpoint-data/stratos-types';
 
-function makeQuota(overrides: Partial<StOrgQuota>): StOrgQuota {
+function makeStack(overrides: Partial<StStack>): StStack {
   return {
-    guid: 'q-1',
-    name: 'default',
-    totalMemoryInMB: 102400,
-    totalInstanceMemoryInMB: 2048,
-    totalInstances: 1000,
-    totalAppTasks: 500,
-    paidServicesAllowed: true,
-    totalServiceInstances: 250,
-    totalServiceKeys: 250,
-    totalRoutes: 1000,
-    totalReservedPorts: 100,
-    totalDomains: 10,
-    organizationCount: 2,
+    guid: 'stack-1',
+    name: 'cflinuxfs4',
+    description: 'Cloud Foundry Linux-based filesystem',
+    default: true,
     cnsiGuid: 'cnsi-1',
     createdAt: '2026-04-22T12:00:00Z',
     updatedAt: '2026-04-22T12:00:00Z',
@@ -28,12 +19,12 @@ function makeQuota(overrides: Partial<StOrgQuota>): StOrgQuota {
   };
 }
 
-function makeHttp(quotas: StOrgQuota[]): HttpClient {
+function makeHttp(stacks: StStack[]): HttpClient {
   return {
     get: vi.fn(() => of({
-      resources: quotas,
+      resources: stacks,
       pagination: {
-        totalResults: quotas.length,
+        totalResults: stacks.length,
         totalPages: 1,
         next: null,
         previous: null,
@@ -44,26 +35,26 @@ function makeHttp(quotas: StOrgQuota[]): HttpClient {
   } as unknown as HttpClient;
 }
 
-function makeSvc(http: HttpClient): CfOrgQuotasSignalConfigService {
+function makeSvc(http: HttpClient): CfStacksSignalConfigService {
   TestBed.configureTestingModule({
     providers: [
       provideZonelessChangeDetection(),
       { provide: HttpClient, useValue: http },
-      CfOrgQuotasSignalConfigService,
+      CfStacksSignalConfigService,
     ],
   });
-  return TestBed.inject(CfOrgQuotasSignalConfigService);
+  return TestBed.inject(CfStacksSignalConfigService);
 }
 
 beforeEach(() => TestBed.resetTestingModule());
 
-describe('CfOrgQuotasSignalConfigService', () => {
-  it('exposes empty quotas before initialize', () => {
+describe('CfStacksSignalConfigService', () => {
+  it('exposes empty stacks before initialize', () => {
     const svc = makeSvc(makeHttp([]));
-    expect(svc.orgQuotas()).toEqual([]);
+    expect(svc.stacks()).toEqual([]);
   });
 
-  it('exposes filter, sort, pageSize, pageIndex, nameFilter signals', () => {
+  it('exposes the filter, sort, pageSize, pageIndex, nameFilter signals', () => {
     const svc = makeSvc(makeHttp([]));
     svc.initialize('cnsi-1');
     expect(svc.filter).toBeDefined();
@@ -73,25 +64,25 @@ describe('CfOrgQuotasSignalConfigService', () => {
     expect(svc.nameFilter).toBeDefined();
   });
 
-  it('builds a ViewPipeline driven by the orgQuotas signal', () => {
+  it('builds a ViewPipeline driven by the stacks signal', () => {
     const svc = makeSvc(makeHttp([]));
     svc.initialize('cnsi-1');
     expect(svc.view).toBeDefined();
   });
 
-  it('loads org quotas from /pp/v1/cf/organization_quotas/:cnsi', async () => {
+  it('loads stacks from /pp/v1/cf/stacks/:cnsi', async () => {
     const http = makeHttp([
-      makeQuota({ guid: 'q-1', name: 'default' }),
-      makeQuota({ guid: 'q-2', name: 'large' }),
+      makeStack({ guid: 'stack-1', name: 'cflinuxfs4' }),
+      makeStack({ guid: 'stack-2', name: 'windows', default: false }),
     ]);
     const svc = makeSvc(http);
     svc.initialize('cnsi-1');
     await svc.loadAll();
 
-    expect(svc.orgQuotas().length).toBe(2);
-    expect(svc.orgQuotas()[0].name).toBe('default');
+    expect(svc.stacks().length).toBe(2);
+    expect(svc.stacks()[0].name).toBe('cflinuxfs4');
     expect(http.get).toHaveBeenCalledWith(
-      expect.stringContaining('/pp/v1/cf/organization_quotas/cnsi-1'),
+      expect.stringContaining('/pp/v1/cf/stacks/cnsi-1'),
     );
   });
