+++ src/frontend/packages/cloud-foundry/src/shared/signal-list-configs/cf-stacks/cf-stacks-signal-config.service.ts
@@ -3,52 +3,54 @@ import { HttpClient } from '@angular/common/http';
 
 import { ListStateStore } from '@stratosui/core';
 
-import { CnsiFeatureFlagsSource } from '../../../../../services/data-sources/cnsi-feature-flags-source';
-import { ViewPipeline, SortSpec } from '../../../../../services/data-sources/view-pipeline';
-import type { StFeatureFlag } from '../../../../../services/endpoint-data/stratos-types';
+import { CnsiStacksSource } from '../../../services/data-sources/cnsi-stacks-source';
+import { ViewPipeline, SortSpec } from '../../../services/data-sources/view-pipeline';
+import type { StStack } from '../../../services/endpoint-data/stratos-types';
 
-// CF Feature Flags list config — single-CNSI, read-only. Drives the
-// per-CF /cloud-foundry/:cnsi/feature-flags tab. Foundations expose
-// ~15 flags so this service stays simple: own the fetch, expose a
-// featureFlags() signal, run it through ViewPipeline. No writes —
-// flag toggle is a platform-admin operation not surfaced here.
+// CF Stacks list config — single-CNSI, read-only. Drives the per-CF
+// /cloud-foundry/:cnsi/stacks tab. Stacks are foundation-level metadata
+// (rootfs flavors); they don't relate to orgs / spaces / apps and are
+// rarely many (typically <10), so this service stays simple: own the
+// fetch, expose a stacks() signal, run it through ViewPipeline. No
+// writes — stack creation/edit/delete are platform-admin operations
+// not surfaced here.
 @Injectable({ providedIn: 'root' })
-export class CfFeatureFlagsSignalConfigService {
+export class CfStacksSignalConfigService {
   private readonly http = inject(HttpClient);
   private readonly injector = inject(Injector);
 
   private cnsiGuid = '';
-  private source?: CnsiFeatureFlagsSource;
+  private source?: CnsiStacksSource;
 
-  private readonly state = inject(ListStateStore).bind('cf-feature-flags', {
+  private readonly state = inject(ListStateStore).bind('cf-stacks', {
     viewMode: 'card',
     pageSize: [24, 25],
     pageIndex: [0, 0],
     sort: [{ field: 'name', direction: 'asc' }, { field: 'name', direction: 'asc' }],
   });
 
-  readonly filter: WritableSignal<(f: StFeatureFlag) => boolean> = signal(() => true);
-  readonly sort = this.state.sort as WritableSignal<SortSpec<StFeatureFlag>>;
+  readonly filter: WritableSignal<(s: StStack) => boolean> = signal(() => true);
+  readonly sort = this.state.sort as WritableSignal<SortSpec<StStack>>;
   readonly pageSize = this.state.pageSize;
   readonly pageIndex = this.state.pageIndex;
   readonly nameFilter: WritableSignal<string> = signal('');
   readonly viewMode = this.state.viewMode;
 
-  private readonly _featureFlags: WritableSignal<StFeatureFlag[]> = signal([]);
-  readonly featureFlags: Signal<StFeatureFlag[]> = this._featureFlags.asReadonly();
+  private readonly _stacks: WritableSignal<StStack[]> = signal([]);
+  readonly stacks: Signal<StStack[]> = this._stacks.asReadonly();
 
   private readonly _hasLoadedOnce: WritableSignal<boolean> = signal(false);
   readonly hasLoadedOnce: Signal<boolean> = this._hasLoadedOnce.asReadonly();
 
-  private readonly _sortExtractors: WritableSignal<Map<string, (row: StFeatureFlag) => unknown>> = signal(new Map());
+  private readonly _sortExtractors: WritableSignal<Map<string, (row: StStack) => unknown>> = signal(new Map());
 
-  view!: ViewPipeline<StFeatureFlag>;
+  view!: ViewPipeline<StStack>;
 
   initialize(cnsiGuid: string): void {
     this.cnsiGuid = cnsiGuid;
-    this.source = new CnsiFeatureFlagsSource(cnsiGuid, this.http);
-    this.view = new ViewPipeline<StFeatureFlag>(
-      this.featureFlags,
+    this.source = new CnsiStacksSource(cnsiGuid, this.http);
+    this.view = new ViewPipeline<StStack>(
+      this.stacks,
       this.filter,
       this.sort,
       this.pageSize,
@@ -59,9 +61,9 @@ export class CfFeatureFlagsSignalConfigService {
     runInInjectionContext(this.injector, () => {
       effect(() => {
         const q = this.nameFilter().trim().toLowerCase();
-        this.filter.set((f: StFeatureFlag) => {
+        this.filter.set((s: StStack) => {
           if (!q) return true;
-          return (f.name ?? '').toLowerCase().includes(q);
+          return (s.name ?? '').toLowerCase().includes(q);
         });
       });
     });
@@ -70,14 +72,14 @@ export class CfFeatureFlagsSignalConfigService {
   async loadAll(): Promise<void> {
     if (!this.source) return;
     await this.source.load();
-    this._featureFlags.set([...this.source.items()]);
+    this._stacks.set([...this.source.items()]);
     this._hasLoadedOnce.set(true);
   }
 
   async refresh(): Promise<void> {
     if (!this.source) return;
     await this.source.refresh();
-    this._featureFlags.set([...this.source.items()]);
+    this._stacks.set([...this.source.items()]);
     this._hasLoadedOnce.set(true);
   }
 
@@ -87,7 +89,7 @@ export class CfFeatureFlagsSignalConfigService {
     this.pageIndex.set(0);
   }
 
-  registerSortExtractor(fieldKey: string, extractor: (row: StFeatureFlag) => unknown): void {
+  registerSortExtractor(fieldKey: string, extractor: (row: StStack) => unknown): void {
     this._sortExtractors.update(curr => {
       const next = new Map(curr);
       next.set(fieldKey, extractor);
